Add better error messages for invalid natural unit
