+++ src/errors/coreErrors.ts
@@ -24,7 +24,8 @@ export const coreAPIErrors = {
   EMPTY_ARRAY: (variable: string) => `The array ${variable} cannot be empty.`,
   EXPIRATION_PASSED: () => 'Expiration date has already passed.',
   FILL_EXCEEDS_AVAILABLE: () => 'The fill quantity supplied exceeds the amount available to fill.',
-  INVALID_NATURAL_UNIT: () => 'Natural unit must be larger than minimum unit allowed by components.',
+  INVALID_NATURAL_UNIT: (minNaturalUnit: BigNumber) =>
+    `Natural unit must be larger than minimum unit, ${minNaturalUnit.toString()} allowed by components.`,
   INVALID_ORDER_HASH_SCHEMA: () => `The order hash passed in is not in Bytes32 schema`,
   QUANTITY_NEEDS_TO_BE_MULTIPLE_OF_NATURAL_UNIT: () => 'Quantity needs to be multiple of natural unit.',
   QUANTITY_NEEDS_TO_BE_POSITIVE: (quantity: BigNumber) =>
